Move helper to another file
Follow up for @bpasero from https://github.com/microsoft/vscode/pull/244026#discussion_r2004102518
