--- src/vs/platform/windows/electron-main/windowUtils.ts
@@ -1,16 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-import electron from 'electron';
-
-/**
- * We have some components like `NativeWebContentExtractorService` that create offscreen windows
- * to extract content from web pages. These windows are not visible to the user and are not
- * considered part of the main application window. This function filters out those offscreen
- * windows from the list of all windows.
- * @returns An array of all BrowserWindow instances that are not offscreen.
- */
-export function getAllWindowsExcludingOffscreen() {
-	return electron.BrowserWindow.getAllWindows().filter(win => !win.webContents.isOffscreen());
-}
